+++ source/Common/Common.csproj
@@ -92,7 +92,10 @@
     <Compile Include="IUpdatable.cs" />
     <Compile Include="LogicStates\ILogic.cs" />
     <Compile Include="LogicStates\IState.cs" />
+    <Compile Include="Messages\IMessageBroker.cs" />
     <Compile Include="Messages\IResponse.cs" />
+    <Compile Include="Messages\MessageBroker.cs" />
+    <Compile Include="Messages\MessagePayload.cs" />
     <Compile Include="MovingAverage.cs" />
     <Compile Include="OrderedHashSet.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
