+++ source/Common/Messages/MessagePayload.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using ProtoBuf;
+
+namespace Common.Messages
+{
+    public readonly struct MessagePayload<T>
+    {
+        public object Who { get; }
+        public T What { get; }
+        public DateTime When { get; }
+
+        public MessagePayload(object who, T what)
+        {
+            Who = who;
+            What = what;
+            When = DateTime.UtcNow;
+        }
+    }
+}
