+++ source/Coop.Tests/packages.config
@@ -2,8 +2,6 @@
 <packages>
   <package id="Autofac" version="6.4.0" targetFramework="net472" />
   <package id="Microsoft.Bcl.AsyncInterfaces" version="1.1.0" targetFramework="net472" />
-  <package id="MSTest.TestAdapter" version="2.2.7" targetFramework="net472" />
-  <package id="MSTest.TestFramework" version="2.2.7" targetFramework="net472" />
   <package id="NLog" version="5.0.1" targetFramework="net472" />
   <package id="System.Buffers" version="4.5.1" targetFramework="net472" />
   <package id="System.Diagnostics.DiagnosticSource" version="4.7.1" targetFramework="net472" />
