+++ source/Coop/Communication/PacketHandlers/PacketManager.cs
@@ -2,8 +2,8 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Common.Messages;
 using Common.Serialization;
-using Coop.Communication.MessageBroker;
 using Coop.Communication.PacketHandlers;
 using Coop.Mod.Messages.Network;
 using GameInterface.Serialization;
