+++ source/Coop/Mod/EventHandlers/EventHandlerBase.cs
@@ -1,5 +1,4 @@
-﻿using Coop.Communication.MessageBroker;
-
+﻿using Common.Messages;
 using GameInterface;
 
 namespace Coop.Mod.EventHandlers
