+++ source/Coop/Mod/Server/States/ServerLogic.cs
@@ -1,5 +1,5 @@
 ﻿using Common.LogicStates;
-using Coop.Communication.MessageBroker;
+using Common.Messages;
 using Coop.Mod.LogicStates.Server;
 
 namespace Coop.Mod.Server.States
