+++ source/GameInterface/GameInterface.cs
@@ -1,9 +1,12 @@
-﻿using GameInterface.Helpers;
+﻿using Common.Messages;
+using GameInterface.Helpers;
 
 namespace GameInterface
 {
     public class GameInterface : IGameInterface
     {
+        public static IMessageBroker MessageBroker { get; }
+
         public IExampleGameHelper ExampleGameHelper { get; }
 
         public ISaveLoadHelper SaveLoadHelper { get; }
